fix: libc fix and timer WIP

+++ src/hal/libc.h
@@ -6,6 +6,6 @@
 #define PERIPHERAL_BASE 0xFE000000
 
 void mmio_write(uint64_t reg, uint32_t val);
-unsigned int mmio_read(uint64_t reg);
+uint64_t mmio_read(uint64_t reg);
 
 #endif
